Split drop OG image route renderer
