+++ app/api/og-metadata/drops/[id]/image/constants.ts
@@ -1,0 +1,171 @@
+import { publicEnv } from "@/config/env";
+
+const CANVAS_WIDTH = 1200;
+const CANVAS_HEIGHT = 630;
+const HORIZONTAL_MARGIN = 46;
+const LOGO_URL = `${publicEnv.BASE_ENDPOINT}/6529.svg`;
+const LOGO_SIZE = 42;
+const CARD_BACKGROUND = "#1A1A1F";
+const CONTENT_BACKGROUND = "#131316";
+const MUTED_TEXT = "#93939F";
+const LINK_TEXT = "#79B8FF";
+const VIDEO_TEXT = "rgba(167, 139, 250, 0.9)";
+const VIDEO_ICON_TEXT = "rgba(167, 139, 250, 0.78)";
+const VIDEO_ICON_BACKGROUND = "rgba(167, 139, 250, 0.08)";
+const VIDEO_ICON_BORDER = "rgba(167, 139, 250, 0.25)";
+const FILE_TEXT = "#818CF8";
+const FILE_ICON_TEXT = "#818CF8";
+const FILE_ICON_BACKGROUND = "rgba(129, 140, 248, 0.08)";
+const FILE_ICON_BORDER = "rgba(129, 140, 248, 0.25)";
+const AUTHOR_ROW_TOP = 44;
+const AUTHOR_AVATAR_SIZE = 70;
+const AUTHOR_AVATAR_INNER_SIZE = 64;
+const AUTHOR_BADGE_SIZE = 38;
+const CONTENT_TOP = 146;
+const CONTENT_FONT_SIZE = 42;
+const CONTENT_LINE_HEIGHT = 1.2;
+const CONTENT_MAX_LINES_WITH_MEDIA = 5;
+const CONTENT_MAX_LINES_WITHOUT_MEDIA = 8;
+const LONG_TEXT_MEDIA_THRESHOLD_LINES = 5;
+const CONTENT_WIDTH = CANVAS_WIDTH - HORIZONTAL_MARGIN * 2;
+const MEDIA_GAP = 24;
+const MEDIA_BOTTOM_MARGIN = 42;
+const MEDIA_MIN_HEIGHT = 140;
+const MEDIA_GALLERY_GAP = 12;
+const MEDIA_GALLERY_MAX_ITEMS = 4;
+const ATTACHMENT_ROWS_WITH_MEDIA = 2;
+const ATTACHMENT_ROWS_WITHOUT_MEDIA = 3;
+const ATTACHMENT_GROUP_TOP_MARGIN = 18;
+const ATTACHMENT_ROW_TOP_MARGIN = 8;
+const ATTACHMENT_ROW_BOTTOM_MARGIN = 6;
+const LINK_DISPLAY_MAX_LINES = 2;
+const LINK_ICON = "🔗";
+const SUBMISSION_TITLE_TOP = 138;
+const SUBMISSION_TITLE_FONT_SIZE = 40;
+const SUBMISSION_TITLE_ICON_SIZE = 40;
+const SUBMISSION_TROPHY_ICON_SIZE = 30;
+const SUBMISSION_TITLE_GAP = 12;
+const ADDITIONAL_ACTION_PROMISE_LABEL = "Additional Action";
+const ADDITIONAL_ACTION_PROMISE_TITLE_BADGE_WIDTH = 156;
+const ADDITIONAL_ACTION_PROMISE_TITLE_BADGE_HEIGHT = 24;
+const ADDITIONAL_ACTION_PROMISE_TITLE_FONT_SIZE = 12;
+const ADDITIONAL_ACTION_PROMISE_CONTEXT_FONT_SIZE = 22;
+const ADDITIONAL_ACTION_PROMISE_CONTEXT_GAP = 12;
+const SUBMISSION_MEDIA_TOP = 206;
+const SUBMISSION_MEDIA_HEIGHT = 344;
+const SUBMISSION_MEDIA_VERTICAL_PADDING = 18;
+const SUBMISSION_MEDIA_CONTENT_HEIGHT =
+  SUBMISSION_MEDIA_HEIGHT - SUBMISSION_MEDIA_VERTICAL_PADDING * 2;
+const SUBMISSION_STATS_TOP = 566;
+const SUBMISSION_STATS_FONT_SIZE = 24;
+const SUBMISSION_STATS_GROUP_GAP = 24;
+const SUBMISSION_STATS_VALUE_GAP = 8;
+const CHAT_CONTENT_BACKGROUND_TOP = CONTENT_TOP - 20;
+const SUBMISSION_STATS_TROPHY_ICON_SIZE = 22;
+const ADDITIONAL_ACTION_PROMISE_TEXT = "#FCD34D";
+const ADDITIONAL_ACTION_PROMISE_BACKGROUND = "rgba(251, 191, 36, 0.1)";
+const ADDITIONAL_ACTION_PROMISE_BORDER = "rgba(252, 211, 77, 0.25)";
+const TROPHY_COLOR = "#FBBF24";
+const SUBMISSION_DATE_FORMATTER = new Intl.DateTimeFormat("en-US", {
+  day: "numeric",
+  month: "short",
+  timeZone: "UTC",
+  year: "numeric",
+});
+const URL_TOKEN_PATTERN = /^(?:https?:\/\/|www\.)\S+$/i;
+const CONTENT_REFERENCE_PATTERN = /([@#])\[([^\]\r\n]+)\]/g;
+const IMAGE_URL_PATTERN = /\.(?:avif|gif|jpe?g|png|webp)(?:$|[?#])/i;
+const INTERACTIVE_URL_PATTERN = /\.(?:glb|gltf|html?)(?:$|[?#])/i;
+const VIDEO_URL_PATTERN = /\.(?:m4v|mov|mp4|webm)(?:$|[?#])/i;
+const UUID_FILENAME_PATTERN =
+  /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+const HEX_FILENAME_PATTERN = /^[0-9a-f]{24,}$/i;
+const NUMERIC_FILENAME_PATTERN = /^\d{10,}$/;
+const GENERIC_FILENAME_TOKENS = new Set([
+  "asset",
+  "file",
+  "image",
+  "img",
+  "media",
+  "upload",
+  "untitled",
+  "vid",
+  "video",
+]);
+
+export {
+  ADDITIONAL_ACTION_PROMISE_BACKGROUND,
+  ADDITIONAL_ACTION_PROMISE_BORDER,
+  ADDITIONAL_ACTION_PROMISE_CONTEXT_FONT_SIZE,
+  ADDITIONAL_ACTION_PROMISE_CONTEXT_GAP,
+  ADDITIONAL_ACTION_PROMISE_LABEL,
+  ADDITIONAL_ACTION_PROMISE_TEXT,
+  ADDITIONAL_ACTION_PROMISE_TITLE_BADGE_HEIGHT,
+  ADDITIONAL_ACTION_PROMISE_TITLE_BADGE_WIDTH,
+  ADDITIONAL_ACTION_PROMISE_TITLE_FONT_SIZE,
+  ATTACHMENT_GROUP_TOP_MARGIN,
+  ATTACHMENT_ROW_BOTTOM_MARGIN,
+  ATTACHMENT_ROW_TOP_MARGIN,
+  ATTACHMENT_ROWS_WITH_MEDIA,
+  ATTACHMENT_ROWS_WITHOUT_MEDIA,
+  AUTHOR_AVATAR_INNER_SIZE,
+  AUTHOR_AVATAR_SIZE,
+  AUTHOR_BADGE_SIZE,
+  AUTHOR_ROW_TOP,
+  CANVAS_HEIGHT,
+  CANVAS_WIDTH,
+  CARD_BACKGROUND,
+  CHAT_CONTENT_BACKGROUND_TOP,
+  CONTENT_BACKGROUND,
+  CONTENT_FONT_SIZE,
+  CONTENT_LINE_HEIGHT,
+  CONTENT_MAX_LINES_WITH_MEDIA,
+  CONTENT_MAX_LINES_WITHOUT_MEDIA,
+  CONTENT_REFERENCE_PATTERN,
+  CONTENT_TOP,
+  CONTENT_WIDTH,
+  FILE_ICON_BACKGROUND,
+  FILE_ICON_BORDER,
+  FILE_ICON_TEXT,
+  FILE_TEXT,
+  GENERIC_FILENAME_TOKENS,
+  HEX_FILENAME_PATTERN,
+  HORIZONTAL_MARGIN,
+  IMAGE_URL_PATTERN,
+  INTERACTIVE_URL_PATTERN,
+  LINK_DISPLAY_MAX_LINES,
+  LINK_ICON,
+  LINK_TEXT,
+  LOGO_SIZE,
+  LOGO_URL,
+  LONG_TEXT_MEDIA_THRESHOLD_LINES,
+  MEDIA_BOTTOM_MARGIN,
+  MEDIA_GALLERY_GAP,
+  MEDIA_GALLERY_MAX_ITEMS,
+  MEDIA_GAP,
+  MEDIA_MIN_HEIGHT,
+  MUTED_TEXT,
+  NUMERIC_FILENAME_PATTERN,
+  SUBMISSION_DATE_FORMATTER,
+  SUBMISSION_MEDIA_CONTENT_HEIGHT,
+  SUBMISSION_MEDIA_HEIGHT,
+  SUBMISSION_MEDIA_TOP,
+  SUBMISSION_STATS_FONT_SIZE,
+  SUBMISSION_STATS_GROUP_GAP,
+  SUBMISSION_STATS_TOP,
+  SUBMISSION_STATS_TROPHY_ICON_SIZE,
+  SUBMISSION_STATS_VALUE_GAP,
+  SUBMISSION_TITLE_FONT_SIZE,
+  SUBMISSION_TITLE_GAP,
+  SUBMISSION_TITLE_ICON_SIZE,
+  SUBMISSION_TITLE_TOP,
+  SUBMISSION_TROPHY_ICON_SIZE,
+  TROPHY_COLOR,
+  URL_TOKEN_PATTERN,
+  UUID_FILENAME_PATTERN,
+  VIDEO_ICON_BACKGROUND,
+  VIDEO_ICON_BORDER,
+  VIDEO_ICON_TEXT,
+  VIDEO_TEXT,
+  VIDEO_URL_PATTERN,
+};
